+++ iterator.go
@@ -7,12 +7,12 @@ type Iterator interface {
 	Next() (ok bool)
 	Prev() (ok bool)
 	Val() []byte
-	Key() int
+	Key() int64
 }
 
 type iterable struct {
 	curr *Node
-	key  int
+	key  int64
 	val  []byte
 }
 
@@ -41,7 +41,7 @@ func (i *iterable) Val() []byte {
 	return i.curr.val
 }
 
-func (i *iterable) Key() int {
+func (i *iterable) Key() int64 {
 	if i.curr == nil {
 		return 0
 	}
